refactor: move to userBalance composable
